remove versioning/add local file reference
Storing the version in a script folder name was going to create messy
diffs and I plan to use a different method to version this project once
converted to modular format.

Adding the local external file reference allows for easily creating a DDR
without first moving the hosted file to a server. I also plan to allow for
all testing to be done without a server in a future release.